fix(ios/android): add auto_route, rootNavigatorKey, fix all compilation errors
- Add auto_route ^9.2.0 to pubspec.yaml (root cause of most errors)
- Create routes.dart defining rootNavigatorKey (GlobalKey<NavigatorState>)
- Update music_app_router.dart to pass rootNavigatorKey to RootStackRouter super()
- Fix desktop_webview_window stub: add CreateConfiguration, setOnUrlRequestCallback,
  clearAll(), update create() signature, remove Cookie class (avoids ambiguity
  with flutter_inappwebview), getAllCookies() returns List<dynamic>
- Migrate SleepTimerNotifier from StateNotifier to Notifier<Duration?> (riverpod 3.x)
- Fix height.ceil() type error in player.dart and player_overlay_collapsed.dart
- Fix sourcedTrack return type cast in playback.dart
- Fix switch exhaustiveness: cast loopMode as PlaylistMode in control.dart
- Fix switch exhaustiveness: cast arg as HistoryDuration in albums.dart and tracks.dart

+++ lib/modules/music/modules/player/player.dart
@@ -48,7 +48,7 @@ class PlayerView extends HookConsumerWidget {
     final shouldHide = useState(true);
 
     ref.listen(navigationPanelHeight, (_, height) {
-      shouldHide.value = height.ceil() == 50;
+      shouldHide.value = (height as double).ceil() == 50;
     });
 
     if (shouldHide.value) {
